pyproject.toml
[tool.poetry]
name = "pinbot"
version = "2.1.11"
description = "Pinbot!"
authors = ["Zenith <z@zenith.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
aurflux = "*"
"discord.py" = "^1.4.1"

[tool.poetry.dev-dependencies]
mypy = "^0.790"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
